refactor(memory): remove obsolete readonly recovery
